fix: claimed and update polkadot api (#104)
* new fetch event from block

* handle every connection exit and reconnect after connection exit

* fix handler event from block

* remove hard code

* remove unused log

* update service request and polkadot version

* integration test service request claimed

* update code in if statement to function.

* docs: add codeowner

+++ src/substrate/substrate.handler.ts
@@ -179,9 +179,7 @@ const eventRoutes = {
 };
 
 @Injectable()
-export class SubstrateService
-  implements OnModuleInit, OnModuleDestroy, OnApplicationBootstrap
-{
+export class SubstrateService implements OnModuleInit, OnModuleDestroy {
   private listenStatus = false;
   private api: ApiPromise;
   private lastBlockNumber = 0;
@@ -190,6 +188,7 @@ export class SubstrateService
   private isFetching = false;
   private fetchBlockInterval: NodeJS.Timer;
   private readonly logger: Logger = new Logger(SubstrateService.name);
+  private isError = false;
   constructor(
     private commandBus: CommandBus,
     private queryBus: QueryBus,
@@ -198,12 +197,22 @@ export class SubstrateService
     private readonly schedulerRegistry: SchedulerRegistry,
   ) {}
 
-  async onApplicationBootstrap() {
+  async onModuleInit() {
     await this.initializeIndices();
+    await this.setWsProvider();
     await this.startListen();
   }
 
-  async onModuleInit() {
+  onModuleDestroy() {
+    this.stopListen();
+  }
+
+  async setWsProvider() {
+    if (this.wsProvider) {
+      this.wsProvider.disconnect();
+      delete this.wsProvider;
+    }
+
     this.wsProvider = new WsProvider(this.process.env.SUBSTRATE_URL);
 
     this.wsProvider.on('connected', () => {
@@ -219,10 +228,6 @@ export class SubstrateService
       this.logger.log(`WS Error: ${error}`);
       await this.restartConnection();
     });
-  }
-
-  onModuleDestroy() {
-    this.stopListen();
   }
 
   async handleEvent(blockMetaData: BlockMetaData, event: Event) {
@@ -251,37 +256,37 @@ export class SubstrateService
   async eventFromBlock(blockNumber: number, blockHash: string | Uint8Array) {
     await this.updateMetaData(blockHash);
 
-    const apiAt = await this.api.at(blockHash);
+    const signedBlock = await this.api.rpc.chain.getBlock(blockHash);
 
-    const allEventsFromBlock = await apiAt.query.system.events();
-
-    const events = allEventsFromBlock.filter(
-      ({ phase }) => phase.isApplyExtrinsic,
-    );
+    const apiAt = await this.api.at(signedBlock.block.header.hash);
 
     const blockMetaData: BlockMetaData = {
       blockNumber: blockNumber,
       blockHash: blockHash.toString(),
     };
 
-    for (let i = 0; i < events.length; i++) {
-      const { event } = events[i];
-      await this.handleEvent(blockMetaData, event);
-    }
+    const allRecords = (await apiAt.query.system.events()) as any;
+
+    signedBlock.block.extrinsics.forEach(
+      // eslint-disable-next-line
+      async ({ method: { method, section } }, index) => {
+        // filter the specific events based on the phase and then the
+        // index of our extrinsic in the block
+        const events = allRecords.filter(
+          ({ phase }) =>
+            phase.isApplyExtrinsic && phase.asApplyExtrinsic.eq(index),
+        );
+        for (let i = 0; i < events.length; i++) {
+          const { event } = events[i];
+          await this.handleEvent(blockMetaData, event);
+        }
+      },
+    );
   }
 
   async listenNewBlocks() {
+    if (this.isError) return;
     try {
-      /**
-       * when this function is called
-       * first check connection to node
-       * restart connection when not connected
-       */
-      if (!this.api.isConnected) {
-        await this.restartConnection();
-        return;
-      }
-
       const currentBlock = await this.api.rpc.chain.getBlock();
       const currentBlockNumber =
         currentBlock.block.header.number.toNumber() - 1;
@@ -315,7 +320,9 @@ export class SubstrateService
         }
       }
     } catch (err) {
+      this.isError = true;
       this.logger.log(`Catch error: ${err.name}, ${err.message}, ${err.stack}`);
+      await this.restartConnection();
     } finally {
       this.isFetching = false;
     }
@@ -391,6 +398,7 @@ export class SubstrateService
       this.logger.log(
         `Handling sync block catch : ${err.name}, ${err.message}, ${err.stack}`,
       );
+      await this.restartConnection();
     }
   }
 
@@ -436,30 +444,33 @@ export class SubstrateService
     this.fetchBlockInterval = setInterval(async () => {
       await this.listenNewBlocks();
     }, 6000);
-
     this.schedulerRegistry.addInterval('fetch-block', this.fetchBlockInterval);
+
+    this.isError = false;
   }
 
   stopListen() {
     this.listenStatus = false;
 
     if (this.api) {
+      this.api.disconnect();
       delete this.api;
     }
 
-    this.schedulerRegistry.deleteInterval('fetch-block');
-    clearInterval(this.fetchBlockInterval);
+    if (this.schedulerRegistry.doesExist('interval', 'fetch-block')) {
+      this.schedulerRegistry.deleteInterval('fetch-block');
+      clearInterval(this.fetchBlockInterval);
+    }
   }
 
   /**
-   * If the connection is active, stop it, then start it again
+   * It stops listening to the node, sets the websocket provider, and starts listening to the node.
    */
   async restartConnection() {
-    if (this.listenStatus) {
-      this.logger.log('Restart connection to node.');
-      this.stopListen();
-      await this.startListen();
-    }
+    this.logger.log('Restart connection to node.');
+    this.stopListen();
+    await this.setWsProvider();
+    await this.startListen();
   }
 
   /**
@@ -501,6 +512,7 @@ export class SubstrateService
         'data-bounty',
         'services',
         'orders',
+        'last-block-number-substrate',
       ];
 
       for (const i of indices) {
